requirements.txt
flask==3.0.1
Werkzeug==3.0.1
jinja2==3.1.3
pure-python-adb==0.3.0.dev0
Flask-SocketIO==5.3.6
protobuf>=4.24.2
pyaxmlparser==0.3.30
googleapis-common-protos==1.62.0
up-transport-zenoh-python@git+https://github.com/eclipse-uprotocol/up-transport-zenoh-python.git@main
vsomeip_py @git+https://github.com/neelam-kushwah/vsomeip_py.git@832d060235663f9d56efc8cd65a313379326cec8
